Model refactorization.

+++ backend/src/main/java/com/example/backend/repository/CertificationRepostory.java
@@ -3,5 +3,5 @@ package com.example.backend.repository;
 import com.example.backend.model.Certificate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CertificationRepostory extends JpaRepository<Certificate,Long> {
+public interface CertificationRepostory extends JpaRepository<Certificate, Integer> {
 }
